@@ -1,43 +1,54 @@
 {
-  "name": "myforum",
-  "version": "0.0.0",
-  "main": "index.js",
-  "dependencies": {
-    "jquery": "^3.4.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "autoprefixer": "^9.5.1",
-    "babel-loader": "^8.0.5",
-    "browser-sync": "^2.26.7",
-    "browser-sync-webpack-plugin": "^2.2.2",
-    "copy-webpack-plugin": "^5.0.2",
-    "css-loader": "^2.1.1",
-    "file-loader": "^3.0.1",
-    "gsap": "^2.1.2",
-    "intersection-observer": "^0.5.1",
-    "jquery-slim": "^3.0.0",
-    "mini-css-extract-plugin": "^0.6.0",
-    "node-sass": "^4.11.0",
-    "path": "^0.12.7",
-    "postcss-loader": "^3.0.0",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.0",
-    "webpack-dev-server": "^3.3.1",
-    "wp-cli": "0.0.5"
-  },
-  "scripts": {
-    "start": "webpack-dev-server --config ./webpack.config.js --mode development",
-    "build": "webpack --config ./webpack.config.js --mode production"
-  },
-  "author": "Stéréosuper",
-  "license": "MIT",
-  "babel": {
-    "presets": [
-      "@babel/preset-env"
-    ]
-  }
+    "name": "myforum",
+    "version": "0.0.0",
+    "main": "index.js",
+    "dependencies": {
+        "@stereorepo/accordion": "^1.3.3",
+        "jquery": "^3.4.1"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.5.4",
+        "@babel/preset-env": "^7.5.4",
+        "autoprefixer": "^9.6.1",
+        "babel-eslint": "^10.0.2",
+        "babel-loader": "^8.0.6",
+        "browser-sync": "^2.26.7",
+        "browser-sync-webpack-plugin": "^2.2.2",
+        "copy-webpack-plugin": "^5.0.3",
+        "css-loader": "^3.0.0",
+        "eslint": "^6.0.1",
+        "eslint-config-prettier": "^6.0.0",
+        "eslint-loader": "^2.2.1",
+        "eslint-plugin-prettier": "^3.1.0",
+        "eslint-plugin-vue": "^5.2.3",
+        "file-loader": "^4.0.0",
+        "gsap": "^2.1.3",
+        "intersection-observer": "^0.7.0",
+        "jquery-slim": "^3.0.0",
+        "mini-css-extract-plugin": "^0.8.0",
+        "node-sass": "^4.12.0",
+        "path": "^0.12.7",
+        "postcss-import": "^12.0.1",
+        "postcss-loader": "^3.0.0",
+        "postcss-nested": "^4.1.2",
+        "postcss-preset-env": "^6.7.0",
+        "prettier": "^1.18.2",
+        "sass-loader": "^7.1.0",
+        "style-loader": "^0.23.1",
+        "webpack": "^4.35.3",
+        "webpack-cli": "^3.3.6",
+        "webpack-dev-server": "^3.7.2",
+        "wp-cli": "0.0.5"
+    },
+    "scripts": {
+        "dev": "webpack-dev-server --config ./webpack.config.js --mode development",
+        "build": "webpack --config ./webpack.config.js --mode production"
+    },
+    "author": "Stéréosuper",
+    "license": "MIT",
+    "babel": {
+        "presets": [
+            "@babel/preset-env"
+        ]
+    }
 }
