@@ -1,13 +1,11 @@
-const $ = require('jquery');
+const $ = require('jquery-slim');
 
 const rotate = () => {
     const blockFatherArrow = $('.subaccordion-title-check');
     const arrow = $('.icon-arrowright');
-    
-    blockFatherArrow.on('click', function(){
+
+    blockFatherArrow.on('click', () => {
         arrow.style.transform = 'rotate90deg';
-    })
-    
-}
-console.log(rotate);
+    });
+};
 export default rotate;
